JobDesignation: fix error on delete item with Employees related

+++ app/src/Backoffice/JobDesignation/Application/Delete/JobDesignationDeleter.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Backoffice\JobDesignation\Application\Delete;
 
 use App\Backoffice\JobDesignation\Application\Get\Single\JobDesignationFinder;
+use App\Backoffice\JobDesignation\Domain\Exception\CannotDeleteJobDesignationWithRelatedEmployees;
 use App\Backoffice\JobDesignation\Domain\JobDesignationRepository;
 
 final class JobDesignationDeleter
@@ -22,6 +23,10 @@ final class JobDesignationDeleter
     public function __invoke(string $id)
     {
         $jobDesignation = $this->finder->__invoke($id);
+
+        if ($jobDesignation->hasEmployees()) {
+            throw new CannotDeleteJobDesignationWithRelatedEmployees();
+        }
 
         $this->repository->delete($jobDesignation);
     }
